server/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
aniso8601==10.0.1
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachecontrol==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloud-sql-python-connector==1.18.3
cloudinary==1.44.1
cryptography==45.0.5
dnspython==2.7.0
faker==37.4.2
firebase-admin==7.0.0
flask==3.1.1
flask-cors==6.0.1
flask-jwt-extended==4.7.1
flask-mail==0.10.0
flask-migrate==4.1.0
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
markupsafe==3.0.2
msgpack==1.1.1
multidict==6.6.3
packaging==25.0
pg8000==1.31.4
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
requests==2.32.4
rsa==4.9.1
scramp==1.4.6
setuptools==70.3.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.29
sqlalchemy-serializer==1.4.22
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
werkzeug==3.1.3
yarl==1.20.1

# AI Agent Dependencies
google-generativeai==0.8.3
PyPDF2==3.0.1
